+++ src/dp/builder.rs
@@ -180,14 +180,6 @@ impl DynamicProgramBuilder {
         self
     }
 
-    /// Sets the [`Kernel`s](crate::kernel::Kernel) for the dynamic program. Use this in combination
-    /// with a [`MultiDynamicProgram`].
-    // pub fn kernels(mut self, kernels: Vec<Kernel>) -> Self {
-    //     self.kernels = Some(kernels);
-    //
-    //     self
-    // }
-
     pub fn field_types(mut self, types: Vec<Vec<usize>>) -> Self {
         self.field_types = Some(types);